demo
This is a demo, where I have used pretend values and functions for N,a,b,c,d and zeroes inorder to see if what i have done works. I also used the pcolor function just to see if the matrix A does what it should. I'll upload my edit to the main file to this branch that I think will be able to fit in the overall code... If I am correct in knowing the inputs used for tank 4