Add files via upload
so I ran the timer again with the subtraction of matrices A. I feel it took A LOT longer this time the last with function m started 1 am and finished just now 11 (in this version the plot ends at more then 4000 and the other version without the matrices ends at 400!). The results also look different. The order of the methods are consistent through all functions this time (although not in the order I was hoping for). 

Now I do not know which to choose from I can upload the version without the finding zeroes in the matrices so we can compare.. Something just does not seem right and I'm am not sure if its something with the computer causing the delay or the code so maybe one of you could have a look and give it a try?